+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Added custom 500 error page to catch internal server errors.
 * Re-trigger FileWatchdog whenever `Path to Trace Directory` gets updated on settings
   page in order to automatically import files from newly given path.
+### Fixed
+* Prevent `unique constraint` sql error when importing activity files with same
+  checksum but different name. Note, this happened only in rare situations, when
+  initially importing multiple files, which are not known to workoutizer.
 
 ## [0.17.1](https://github.com/fgebhart/workoutizer/releases/tag/0.17.1) - 2021-04-20
 ### Changed